refator: move graph api client into services
